Updated thick box CSS, added pull remote form

+++ templates/branches.php
@@ -46,7 +46,7 @@
 						$active_tab = isset( $_GET[ 'tab' ] ) ? $_GET[ 'tab' ] : 'local_branches';
 
 						if ( $active_tab == 'remote_branches' ) {
-							$output 	= $git->get_branches( true );
+							$output = $git->get_branches( true );
 						} else {
 							$output = $git->get_branches();
 						}
@@ -61,7 +61,7 @@
 									$num_commits = count( $git->run( "log $branch --pretty=oneline" ) );
 								}
 								
-								if (substr( $value, 0, 1 ) === "*"){
+								if ( substr( $value, 0, 1 ) === "*" ){
 									echo "<tr>
 									<td><strong>$branch (current branch)</strong></td>
 									<td class='center-td'>$num_commits</td>
@@ -71,19 +71,34 @@
 										<a class='button disabled branch-btn' onclick='preventDefault()' href='#'>Delete</a>
 									</td></tr>";
 								} else {
-									$checkout_url 	= get_admin_url() . "admin-post.php?action=process_checkout&branch={$branch}";
-									$merge_url 		= get_admin_url() . "admin-post.php?action=merge_branch_form&branch={$branch}&TB_iframe=true&width=350&height=225";
-									$delete_url 	= get_admin_url() . "admin-post.php?action=delete_branch_form&branch={$branch}&TB_iframe=true&width=350&height=150";
+									$checkout_url 		= get_admin_url() . "admin-post.php?action=process_checkout&branch={$branch}";
+									$merge_url 			= get_admin_url() . "admin-post.php?action=merge_branch_form&branch={$branch}&TB_iframe=true&width=350&height=200";
+									$delete_url 		= get_admin_url() . "admin-post.php?action=delete_branch_form&branch={$branch}&TB_iframe=true&width=350&height=200";
+									$pull_remote_url 	= get_admin_url() . "admin-post.php?action=pull_remote_form&remote_branch={$branch}&TB_iframe=true&width=350&height=200";
 									?>
 									<tr>
 									<td><?php echo $branch; ?></td>
 									<td style='text-align:center;'><?php echo $num_commits; ?></td>
-									<td class="center-td">
-										<a class='button branch-btn' href='<?php echo $checkout_url; ?>'><?php _e( 'Checkout', 'revisr' ); ?></a>
-										<a class='button branch-btn merge-btn thickbox' href="<?php echo $merge_url; ?>" title="<?php _e( 'Merge Branch', 'revisr' ); ?>">Merge</a>
-										<a class='button branch-btn delete-branch-btn thickbox' href='<?php echo $delete_url; ?>' title='<?php _e( 'Delete Branch', 'revisr' ); ?>'><?php _e( 'Delete', 'revisr' ); ?></a>
-									</td></tr>
 									<?php
+									if ( $active_tab == 'remote_branches' ) {
+
+										?>
+										<td class="center-td">
+											<a class="button branch-btn thickbox" href="<?php echo $pull_remote_url; ?>" title="<?php _e( 'Pull Remote', 'revisr' ); ?>"><?php _e( 'Pull into...', 'revisr' ); ?></a>
+										</td></tr>
+										<?php
+
+									} else {
+
+										?>
+										<td class="center-td">
+											<a class='button branch-btn' href='<?php echo $checkout_url; ?>'><?php _e( 'Checkout', 'revisr' ); ?></a>
+											<a class='button branch-btn merge-btn thickbox' href="<?php echo $merge_url; ?>" title="<?php _e( 'Merge Branch', 'revisr' ); ?>">Merge</a>
+											<a class='button branch-btn delete-branch-btn thickbox' href='<?php echo $delete_url; ?>' title='<?php _e( 'Delete Branch', 'revisr' ); ?>'><?php _e( 'Delete', 'revisr' ); ?></a>
+										</td></tr>
+										<?php
+
+									}
 								}
 							}										
 						}
